# docs/api-contract.openapi.yaml
...
openapi: 3.1.0
info:
  title: Parlay chat server API
  version: "2026.08.30"
  description: >
    The /api/chat surface served by both the TS server (packages/server, Bun)
    and the Go server (packages/go-server). No authentication; the surface
    trusts the network boundary (local/tailnet only) plus an origin guard for
    cross-origin browser callers. This file is authoritative over
    docs/api-contract.md on conflict.
servers:
  - url: http://localhost:4242
    description: Default dev/production address (PARLAY_SERVER overrides).
paths:
...
  /api/chat/history:
    get:
      operationId: history
      summary: Recent chat history, oldest first
      description: >-
        `limit` bounds how many of the retained messages are returned (the
        server retains 5,000). An absent, non-numeric or non-positive `limit`
        returns the whole retained window — there is no smaller server-side
        default. `channel` is accepted by some clients and ignored.
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
          description: Maximum retained messages to return; 0 or absent means all retained.
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatMessage" }
...
components:
  schemas:
...
    ChatAction:
      type: object
      required: [kind, label]
      properties:
        kind: { enum: [navigate, switch_tab] }
        url: { type: string }
        channel: { type: string }
        label: { type: string }
    ChatMessage:
      type: object
      required: [id, role, ts, text]
      properties:
        id: { type: string }
        role: { enum: [user, agent] }
        ts: { type: string, description: ISO 8601. }
        text: { type: string }
        channel: { type: string, description: Agent id; absent for global/user messages. }
        type: { enum: [alert, action_request, system_update] }
        action: { $ref: "#/components/schemas/ChatAction" }
        source: { type: string, description: "system_update: emitting hook/system label." }
        meta: { type: object }
        images: { type: array, items: { type: string } }
        from: { type: string, description: User-role sender attribution; absent = the captain. }
        received: { type: boolean, description: Runtime-only delivery pip; stripped from disk. }
...